refactor: Move `Encoded` into submodule
The amount of code related to `Encoded` will increase when separate
classes for fragments will be implemented.
