+++ bindings/imap-codec-python/src/lib.rs
@@ -1,6 +1,9 @@
+mod encoded;
+
+use encoded::PyEncoded;
 use imap_codec::{
     decode::{self, Decoder},
-    encode::{Encoded, Encoder},
+    encode::Encoder,
     AuthenticateDataCodec, CommandCodec, GreetingCodec, IdleDoneCodec, ResponseCodec,
 };
 use pyo3::{create_exception, exceptions::PyException, prelude::*, types::PyBytes};
@@ -10,42 +13,6 @@ create_exception!(imap_codec, DecodeError, PyException);
 create_exception!(imap_codec, DecodeFailed, DecodeError);
 create_exception!(imap_codec, DecodeIncomplete, DecodeError);
 create_exception!(imap_codec, DecodeLiteralFound, DecodeError);
-
-/// Wrapper for `Encoded`
-///
-/// This implements a Python iterator over the containing fragments.
-#[derive(Debug, Clone)]
-#[pyclass(name = "Encoded")]
-struct PyEncoded(Option<Encoded>);
-
-#[pymethods]
-impl PyEncoded {
-    /// Initialize iterator
-    fn __iter__(slf: PyRef<'_, Self>) -> PyRef<'_, Self> {
-        slf
-    }
-
-    /// Return next fragment
-    fn __next__(mut slf: PyRefMut<'_, Self>) -> PyResult<Option<Bound<PyAny>>> {
-        let Some(encoded) = &mut slf.0 else {
-            return Ok(None);
-        };
-        Ok(encoded
-            .next()
-            .map(|value| serde_pyobject::to_pyobject(slf.py(), &value))
-            .transpose()?)
-    }
-
-    /// Dump remaining fragment data
-    fn dump(mut slf: PyRefMut<'_, Self>) -> PyResult<Bound<PyBytes>> {
-        let encoded = slf.0.take();
-        let dump = match encoded {
-            Some(encoded) => encoded.dump(),
-            None => Vec::new(),
-        };
-        Ok(PyBytes::new_bound(slf.py(), &dump))
-    }
-}
 
 /// Python class for using `GreetingCodec`
 #[derive(Debug, Clone, PartialEq)]
